+++ src/Symfony/Component/Messenger/Bridge/Doctrine/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 7.1
 ---
 
+ * Add support to pull a batch of messages in one go (`batch_size` option)
  * Use `SKIP LOCKED` in the doctrine transport for MySQL, PostgreSQL and MSSQL
 
 5.1.0